refactor/utils: rename hash extensions module
Since HashMap extensions were added, set is not appropriate in module name anymore.
