+++ src/util/hash_ext.rs
