+++ src/util/mod.rs
@@ -1,6 +1,6 @@
-mod hash_set_ext;
+mod hash_ext;
 
-pub use self::hash_set_ext::*;
+pub use self::hash_ext::*;
 
 #[cfg(test)]
 #[macro_use]
